+++ framework/contracts/account/manager/src/commands.rs
@@ -143,7 +143,9 @@ pub(crate) fn install_modules_internal(
 
     let (infos, init_msgs): (Vec<_>, Vec<_>) =
         modules.into_iter().map(|m| (m.module, m.init_msg)).unzip();
-    let modules = version_control.query_modules_configs(infos, &deps.querier)?;
+    let modules = version_control
+        .query_modules_configs(infos, &deps.querier)
+        .map_err(|error| ManagerError::QueryModulesFailed { error })?;
 
     let mut install_context = Vec::with_capacity(modules.len());
     let mut to_add = Vec::with_capacity(modules.len());
@@ -278,6 +280,7 @@ pub fn create_sub_account(
     base_asset: Option<AssetEntry>,
     namespace: Option<String>,
     install_modules: Vec<ModuleInstallConfig>,
+    account_id: Option<u32>,
 ) -> ManagerResult {
     // only owner can create a subaccount
     assert_admin_right(deps.as_ref(), &msg_info.sender)?;
@@ -294,7 +297,7 @@ pub fn create_sub_account(
         base_asset,
         namespace,
         install_modules,
-        account_id: None,
+        account_id: account_id.map(AccountId::local),
     };
 
     let account_factory_addr = query_module(