+++ framework/contracts/account/manager/tests/subaccount.rs
@@ -23,6 +23,7 @@ fn creating_on_subaccount_should_succeed() -> AResult {
         None,
         None,
         None,
+        None,
         &[],
     )?;
     let sub_accounts = account.manager.sub_account_ids(None, None)?;
@@ -46,6 +47,7 @@ fn updating_on_subaccount_should_succeed() -> AResult {
     account.manager.create_sub_account(
         vec![],
         "My subaccount".to_string(),
+        None,
         None,
         None,
         None,
@@ -78,6 +80,7 @@ fn proxy_updating_on_subaccount_should_succeed() -> AResult {
     account.manager.create_sub_account(
         vec![],
         "My subaccount".to_string(),
+        None,
         None,
         None,
         None,
@@ -116,6 +119,7 @@ fn recursive_updating_on_subaccount_should_succeed() -> AResult {
         None,
         None,
         None,
+        None,
         &[],
     )?;
 
@@ -126,6 +130,7 @@ fn recursive_updating_on_subaccount_should_succeed() -> AResult {
     account_contracts.0.create_sub_account(
         vec![],
         "My subsubaccount".to_string(),
+        None,
         None,
         None,
         None,
@@ -158,6 +163,7 @@ fn installed_app_updating_on_subaccount_should_succeed() -> AResult {
     account.manager.create_sub_account(
         vec![],
         "My subaccount".to_string(),
+        None,
         None,
         None,
         None,
@@ -217,6 +223,7 @@ fn sub_account_move_ownership() -> AResult {
     account.manager.create_sub_account(
         vec![],
         "My subaccount".to_string(),
+        None,
         None,
         None,
         None,
@@ -297,6 +304,7 @@ fn sub_account_move_ownership_to_sub_account() -> AResult {
         None,
         None,
         None,
+        None,
         &[],
     )?;
     let sub_account = AbstractAccount::new(&deployment, AccountId::local(2));
@@ -307,6 +315,7 @@ fn sub_account_move_ownership_to_sub_account() -> AResult {
     new_account.manager.create_sub_account(
         vec![],
         "My second subaccount".to_string(),
+        None,
         None,
         None,
         None,
@@ -373,6 +382,7 @@ fn account_move_ownership_to_falsy_sub_account() -> AResult {
     account.manager.create_sub_account(
         vec![],
         "My subaccount".to_string(),
+        None,
         None,
         None,
         None,
@@ -488,6 +498,7 @@ fn top_level_owner() -> AResult {
         None,
         None,
         None,
+        None,
         &[],
     )?;
     let response = account.manager.sub_account_ids(None, None)?;
@@ -509,6 +520,7 @@ fn cant_renounce_with_sub_accounts() -> AResult {
     account.manager.create_sub_account(
         vec![],
         "My subaccount".to_string(),
+        None,
         None,
         None,
         None,
@@ -537,6 +549,7 @@ fn can_renounce_sub_accounts() -> AResult {
     account.manager.create_sub_account(
         vec![],
         "My subaccount".to_string(),
+        None,
         None,
         None,
         None,