+++ framework/packages/abstract-client/tests/integration.rs
@@ -1178,3 +1178,59 @@ fn authorize_app_on_adapters() -> anyhow::Result<()> {
     assert_eq!(authorized_addrs_resp.addresses, vec![app.address()?]);
     Ok(())
 }
+
+#[test]
+fn create_account_with_expected_account_id() -> anyhow::Result<()> {
+    let chain = Mock::new(&Addr::unchecked(OWNER));
+    let client = AbstractClient::builder(chain).build()?;
+
+    // Check it fails on wrong account_id
+    let next_id = client.next_local_account_id()?;
+    let err = client.account_builder().account_id(10).build().unwrap_err();
+    let AbstractClientError::Interface(abstract_interface::AbstractInterfaceError::Orch(err)) = err
+    else {
+        panic!("Expected cw-orch error")
+    };
+    let err: account_factory::error::AccountFactoryError = err.downcast().unwrap();
+    assert_eq!(
+        err,
+        account_factory::error::AccountFactoryError::ExpectedAccountIdFailed {
+            predicted: AccountId::local(10),
+            actual: AccountId::local(next_id)
+        }
+    );
+
+    // Can create if right id
+    let account = client.account_builder().account_id(next_id).build()?;
+
+    // Check it fails on wrong account_id for sub-accounts
+    let next_id = client.next_local_account_id()?;
+    let err = client
+        .account_builder()
+        .sub_account(&account)
+        .account_id(0)
+        .build()
+        .unwrap_err();
+    let AbstractClientError::Interface(abstract_interface::AbstractInterfaceError::Orch(err)) = err
+    else {
+        panic!("Expected cw-orch error")
+    };
+    let err: account_factory::error::AccountFactoryError = err.downcast().unwrap();
+    assert_eq!(
+        err,
+        account_factory::error::AccountFactoryError::ExpectedAccountIdFailed {
+            predicted: AccountId::local(0),
+            actual: AccountId::local(next_id)
+        }
+    );
+
+    // Can create sub-account if right id
+    let sub_account = client
+        .account_builder()
+        .sub_account(&account)
+        .account_id(next_id)
+        .build()?;
+    let sub_accounts = account.sub_accounts()?;
+    assert_eq!(sub_accounts[0].id()?, sub_account.id()?);
+    Ok(())
+}